third_party/WebKit/LayoutTests/virtual/outofblink-cors/fast/cookies/local-file-can-set-cookies-expected.txt

This test checks that a local file can set cookies. See: rdar://problem/5379090 REGRESSION: Cannot set cookies for local files

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.cookie.indexOf('ppkcookie1=testcookie') is -1
FAIL document.cookie.indexOf('ppkcookie1=testcookie') >= 0 should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
